# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SAKBFANTASY",
        "description": "Api SAKBFANTASY Documentation",
        "contact": {
            "email": "info@sakb.com"
        },
        "license": {
            "name": "Sakb",
            "url": "http://sakb-co.com.sa/"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/fixtures": {
            "get": {
                "tags": [
                    "subeldwry_fixtures"
                ],
                "summary": "get all fixtures",
                "operationId": "fixtures",
                "parameters": [
                    {
                        "name": "type-dev",
                        "in": "header",
                        "description": "ios or android",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "ios"
                        }
                    },
                    {
                        "name": "val-dev",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "sakb"
                        }
                    },
                    {
                        "name": "lang",
                        "in": "header",
                        "description": "default lang is ar ( ar , en)",
                        "schema": {
                            "type": "string",
                            "default": "ar"
                        }
                    },
                    {
                        "name": "num_page",
                        "in": "path",
                        "description": " number of page strat from zero ( 0 )",
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "path",
                        "description": " limit is number fixtures will send in each time default (1)",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "400": {
                        "description": "error"
                    }
                }
            }
        },
...
        "/api/v1/eldwry_locations": {
            "post": {
                "tags": [
                    "game"
                ],
                "summary": "get current eldwry and locations",
                "operationId": "eldwry_locations",
                "parameters": [
                    {
                        "name": "access-token",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "type-dev",
                        "in": "header",
                        "description": "ios or android",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "ios"
                        }
                    },
                    {
                        "name": "val-dev",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "sakb"
                        }
                    },
                    {
                        "name": "lang",
                        "in": "header",
                        "description": "default lang is ar ( ar , en)",
                        "schema": {
                            "type": "string",
                            "default": "ar"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "400": {
                        "description": "error"
                    }
                }
            }
        },
        "/api/v1/players_by_type": {
            "post": {
                "tags": [
                    "game"
                ],
                "summary": "get all players by type",
                "operationId": "players_by_type",
                "parameters": [
                    {
                        "name": "access-token",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "type-dev",
                        "in": "header",
                        "description": "ios or android",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "ios"
                        }
                    },
                    {
                        "name": "val-dev",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "sakb"
                        }
                    },
                    {
                        "name": "lang",
                        "in": "header",
                        "description": "default lang is ar ( ar , en)",
                        "schema": {
                            "type": "string",
                            "default": "ar"
                        }
                    },
                    {
                        "name": "type_key",
                        "in": "query",
                        "description": "get value from api eldwry_locations and default is goalkeeper (goalkeeper,defender,line,attacker)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "goalkeeper"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "400": {
                        "description": "error"
                    }
                }
            }
        },
        "/api/v1/filterPlayer": {
            "post": {
                "tags": [
                    "game"
                ],
                "summary": "get all Players after filter",
                "operationId": "filterPlayer",
                "parameters": [
                    {
                        "name": "access-token",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "type-dev",
                        "in": "header",
                        "description": "ios or android",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "ios"
                        }
                    },
                    {
                        "name": "val-dev",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "sakb"
                        }
                    },
                    {
                        "name": "lang",
                        "in": "header",
                        "description": "default lang is ar ( ar , en)",
                        "schema": {
                            "type": "string",
                            "default": "ar"
                        }
                    },
                    {
                        "name": "link_team",
                        "in": "query",
                        "description": "",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "type_key",
                        "in": "query",
                        "description": "get value from api eldwry_locations and default is goalkeeper (goalkeeper,defender,line,attacker)",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "word_search",
                        "in": "query",
                        "description": "",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "order_play",
                        "in": "query",
                        "description": "heigh_price or low_price or goon or point",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "from_price",
                        "in": "query",
                        "description": "",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "to_price",
                        "in": "query",
                        "description": "",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "loc_player",
                        "in": "query",
                        "description": "",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "400": {
                        "description": "error"
                    }
                }
            }
        },
...
    },
    "security": [
        []
    ]
}